ui/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
websocket-client>=1.6.4
pandas>=2.1.0
plotly>=5.18.0
python-dotenv>=1.0.0
